refactor: Extract comp from GroupSettings
