+++ src/ducks/settings/GroupSettings/RenameGroupForm.jsx
@@ -1,0 +1,81 @@
+import React, { useState, useRef, useCallback } from 'react'
+
+import { useClient } from 'cozy-client'
+import { Media, Img } from 'cozy-ui/transpiled/react/Media'
+import Input from 'cozy-ui/transpiled/react/Input'
+import { useI18n } from 'cozy-ui/transpiled/react/I18n'
+import Button from 'cozy-ui/transpiled/react/Button'
+import Typography from 'cozy-ui/transpiled/react/Typography'
+
+import { useRouter } from 'components/RouterContext'
+import { getGroupLabel, renamedGroup } from 'ducks/groups/helpers'
+import { trackEvent } from 'ducks/tracking/browser'
+import { updateOrCreateGroup } from 'ducks/settings/GroupSettings/helpers'
+
+import styles from 'ducks/settings/GroupsSettings.styl'
+
+const RenameGroupForm = props => {
+  const [modifying, setModifying] = useState(false)
+  const [saving, setSaving] = useState(false)
+  const client = useClient()
+  const router = useRouter()
+  const inputRef = useRef()
+  const { t } = useI18n()
+
+  const { group } = props
+
+  const handleRename = useCallback(() => {
+    setSaving(true)
+    const updatedGroup = renamedGroup(group, inputRef.current.value)
+    return updateOrCreateGroup(client, updatedGroup, router, () => {
+      setSaving(false)
+      setModifying(false)
+      trackEvent({
+        name: 'renommer'
+      })
+    })
+  }, [client, group, router])
+
+  const handleModifyName = useCallback(() => {
+    setModifying(true)
+  }, [setModifying])
+
+  return (
+    <form className={styles.GrpStg__form} onSubmit={e => e.preventDefault()}>
+      <Media>
+        <Img>
+          {!modifying ? (
+            <Typography variant="body1">{getGroupLabel(group, t)}</Typography>
+          ) : (
+            <Input
+              inputRef={inputRef}
+              placeholder={t('Groups.name-placeholder')}
+              autoFocus
+              type="text"
+              defaultValue={getGroupLabel(group, t)}
+            />
+          )}
+        </Img>
+        <Img>
+          {modifying ? (
+            <Button
+              disabled={saving}
+              theme="regular"
+              onClick={handleRename}
+              label={t('Groups.save')}
+              busy={saving}
+            />
+          ) : (
+            <Button
+              theme="text"
+              onClick={handleModifyName}
+              label={t('Groups.rename')}
+            />
+          )}
+        </Img>
+      </Media>
+    </form>
+  )
+}
+
+export default React.memo(RenameGroupForm)
